+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - New `fetchFilesGlob` data source that can load multiple data files at once
+- Event time is now a core part of the datasets
 
 ## [0.4.0] - 2019-11-24
 ### Added